refactor(jruby): move DriverFactory Ext under Ext::Internal::
The Ruby module that prepends helpers onto Java's
`org.neo4j.driver.internal.DriverFactory` belongs under
`Ext::Internal::DriverFactory`, mirroring Java's `internal` package
(same place `Ext::Internal::ClockAdapter`, `Ext::Internal::Cluster::*`,
`Ext::Internal::Summary::*`, etc. already live). Drops the one-off
`Helpers` suffix; the module gets the same name as the Java class it
extends, matching every other Ext module.
